requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.9.1
openai==1.97.1
elasticsearch[async]==8.17.1
PyMuPDF==1.24.11
playwright==1.58.0
httpx==0.28.1
python-multipart==0.0.20
python-dotenv==1.0.1
